requirements.txt
asgiref==3.11.1
certifi==2026.2.25
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
defusedxml==0.7.1
django==6.0.4
django-axes==8.3.1
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.17.1
h11==0.16.0
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
pip==26.0.1
psycopg2-binary==2.9.11
pygments==2.20.0
pysocks==1.7.1
python-decouple==3.8
python-json-logger==4.1.0
redis==7.4.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
ruff==0.15.10
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
